refactor: 테스트에 필요한 부가 클래스를 support 패키지로 이동
