+++ src/test/java/roomescape/controller/AdminReservationControllerTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.ClearDbTest;
 import roomescape.dto.ReservationResult;
+import roomescape.support.ClearDbTest;
 
 import java.time.LocalDate;
 import java.util.List;
