+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -5,10 +5,10 @@ import io.restassured.http.ContentType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.ClearDbTest;
 import roomescape.dto.AvailableDateResult;
 import roomescape.dto.ReservationResult;
 import roomescape.dto.ReservationTimeStatusResult;
+import roomescape.support.ClearDbTest;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
