+++ src/test/java/roomescape/controller/ReservationTimeControllerTest.java
@@ -5,9 +5,9 @@ import io.restassured.http.ContentType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.ClearDbTest;
 import roomescape.dto.ReservationTimeCreateCommand;
 import roomescape.dto.ReservationTimeResult;
+import roomescape.support.ClearDbTest;
 
 import java.time.LocalTime;
 import java.util.HashMap;
